Fix lint packaging and fail closed runtime paths
