+++ tests/test_lint.py
@@ -7,12 +7,11 @@ from pathlib import Path
 import shutil
 import subprocess
 import sys
-import tomllib
 
 import pytest
 import yaml
 
-from manifests.schema import Manifest, load_manifest
+from open_range.manifest_schema import Manifest, load_manifest
 from open_range.lint import lint_file, lint_manifest
 
 
@@ -291,15 +290,12 @@ class TestLintFile:
 
 
 class TestPackagingAndInvocation:
-    def test_pyproject_includes_manifests_package(self):
-        pyproject = tomllib.loads((ROOT / "pyproject.toml").read_text())
-        packages = pyproject["tool"]["setuptools"]["packages"]
-        assert "manifests" in packages
+    def test_open_range_package_contains_manifest_schema(self):
+        assert (ROOT / "src" / "open_range" / "manifest_schema.py").exists()
 
     def test_lint_module_runs_outside_repo_with_packaged_layout(self, tmp_path):
         site_root = tmp_path / "site"
         shutil.copytree(ROOT / "src" / "open_range", site_root / "open_range")
-        shutil.copytree(ROOT / "src" / "manifests", site_root / "manifests")
 
         outside = tmp_path / "outside"
         outside.mkdir()
